+++ app/main.tsx
@@ -5,6 +5,7 @@ import BanzukePage from './banzuke/page'
 import HomePage from './page'
 import TorikumiHubPage from './torikumi/page'
 import TopLevelSlugPage from './TopLevelSlugPage'
+import RikishiProfilePage from './rikishi/[id]/page'
 import { banzukePath, getHubPath, legacyBanzukePath } from './lib/torikumi-routes'
 import './globals.css'
 
@@ -17,6 +18,7 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
         <Route path={legacyBanzukePath} element={<Navigate to={banzukePath} replace />} />
         <Route path={getHubPath('result')} element={<TorikumiHubPage mode="result" />} />
         <Route path={getHubPath('schedule')} element={<TorikumiHubPage mode="schedule" />} />
+        <Route path="/rikishi/:id" element={<RikishiProfilePage />} />
         <Route path="/:slug" element={<TopLevelSlugPage />} />
       </Routes>
     </BrowserRouter>